Refactor device options retrieval in AddEditDeviceForm.
The commit involves a significant refactor in how the AddEditDeviceForm component fetches options for templates, profiles, and extensions. Options are now gathered from 'device.device_options' instead of being passed in as prop. Furthermore, it updates the 'UpdateDeviceRequest' so that it fetches the 'domain_uuid' from the request rather than from the session. These changes help in making the components loosely coupled and easier to test.